+++ include/mujoco/mujoco.h
@@ -671,14 +671,6 @@ MJAPI void mjv_initGeom(mjvGeom* geom, int type, const mjtNum size[3],
 // Set (type, size, pos, mat) for connector-type geom between given points.
 // Assume that mjv_initGeom was already called to set all other properties.
 // Width of mjGEOM_LINE is denominated in pixels.
-// Deprecated: use mjv_connector.
-MJAPI void mjv_makeConnector(mjvGeom* geom, int type, mjtNum width,
-                             mjtNum a0, mjtNum a1, mjtNum a2,
-                             mjtNum b0, mjtNum b1, mjtNum b2);
-
-// Set (type, size, pos, mat) for connector-type geom between given points.
-// Assume that mjv_initGeom was already called to set all other properties.
-// Width of mjGEOM_LINE is denominated in pixels.
 MJAPI void mjv_connector(mjvGeom* geom, int type, mjtNum width,
                          const mjtNum from[3], const mjtNum to[3]);